Fix clippy warnings across the workspace
Mechanical lints: identity ops (1 * w) in test helpers, len() >= 1 ->
!is_empty(), a redundant u32 cast, a redundant struct-update spread, and
a field-reassign test rewritten with struct-update syntax. The core
roll_affix_set 8-arg helper gets a justified allow rather than a
context-struct refactor that would only shift the params out a level.

+++ crates/procgen-viz/src/main.rs
@@ -123,7 +123,7 @@ fn render<W: Write>(
                         if d == UNREACHABLE {
                             paint(out, '?', Color::Wall, args.color)?;
                         } else {
-                            let ch = char::from_digit((d % 10) as u32, 10).unwrap_or('.');
+                            let ch = char::from_digit(d % 10, 10).unwrap_or('.');
                             paint(out, ch, Color::Floor, args.color)?;
                         }
                     } else {
